sales_data_analyst_agent/requirements.txt
google-generativeai
google-adk
pandas
fastapi
uvicorn
